+++ pymc/progress_bar/rich_progress.py
@@ -1,0 +1,374 @@
+#   Copyright 2025 - present The PyMC Developers
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+"""Rich-based progress bar backend for terminal rendering."""
+
+from __future__ import annotations
+
+from collections.abc import Iterable
+from typing import TYPE_CHECKING, Any
+
+from rich.box import SIMPLE_HEAD
+from rich.console import Console
+from rich.progress import (
+    BarColumn,
+    Progress,
+    Task,
+    TextColumn,
+    TimeElapsedColumn,
+    TimeRemainingColumn,
+)
+from rich.style import Style
+from rich.table import Column, Table
+from rich.theme import Theme
+
+from pymc.progress_bar.progress import compute_draw_speed, default_progress_theme
+
+if TYPE_CHECKING:
+    from rich.progress import TaskID
+
+
+class CustomProgress(Progress):
+    """A child of Progress that allows disabling progress bars and their container.
+
+    The implementation checks an `is_enabled` flag and generates the progress bar
+    only if it's True.
+    """
+
+    def __init__(self, *args, disable: bool = False, include_headers: bool = False, **kwargs):
+        self.is_enabled = not disable
+        self.include_headers = include_headers
+
+        if self.is_enabled:
+            super().__init__(*args, **kwargs)
+
+    def __enter__(self):
+        """Enter the context manager."""
+        if self.is_enabled:
+            self.start()
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        """Exit the context manager."""
+        if self.is_enabled:
+            super().__exit__(exc_type, exc_val, exc_tb)
+
+    def add_task(self, *args, **kwargs) -> TaskID | None:
+        if self.is_enabled:
+            return super().add_task(*args, **kwargs)
+        return None
+
+    def advance(self, task_id, advance=1) -> None:
+        if self.is_enabled:
+            super().advance(task_id, advance)
+        return None
+
+    def update(
+        self,
+        task_id,
+        *,
+        total=None,
+        completed=None,
+        advance=None,
+        description=None,
+        visible=None,
+        refresh=False,
+        **fields,
+    ):
+        if self.is_enabled:
+            super().update(
+                task_id,
+                total=total,
+                completed=completed,
+                advance=advance,
+                description=description,
+                visible=visible,
+                refresh=refresh,
+                **fields,
+            )
+        return None
+
+    def make_tasks_table(self, tasks: Iterable[Task]) -> Table:
+        """Get a table to render the Progress display.
+
+        Unlike the parent method, this returns a full table (not a grid),
+        allowing for column headings.
+
+        Parameters
+        ----------
+        tasks : Iterable[Task]
+            An iterable of Task instances, one per row of the table.
+
+        Returns
+        -------
+        Table
+            A table instance.
+        """
+
+        def call_column(column, task):
+            # Subclass rich.BarColumn and add a callback method to dynamically update display
+            if hasattr(column, "callbacks"):
+                column.callbacks(task)
+            return column(task)
+
+        table_columns = (
+            (
+                Column(no_wrap=True)
+                if isinstance(_column, str)
+                else _column.get_table_column().copy()
+            )
+            for _column in self.columns
+        )
+        if self.include_headers:
+            table = Table(
+                *table_columns,
+                padding=(0, 1),
+                expand=self.expand,
+                show_header=True,
+                show_edge=True,
+                box=SIMPLE_HEAD,
+            )
+        else:
+            table = Table.grid(*table_columns, padding=(0, 1), expand=self.expand)
+
+        for task in tasks:
+            if task.visible:
+                table.add_row(
+                    *(
+                        (
+                            column.format(task=task)
+                            if isinstance(column, str)
+                            else call_column(column, task)
+                        )
+                        for column in self.columns
+                    )
+                )
+
+        return table
+
+
+class RecolorOnFailureBarColumn(BarColumn):
+    """Rich colorbar that changes color when a chain has detected a failure."""
+
+    def __init__(self, *args, failing_color: str = "red", **kwargs):
+        from matplotlib.colors import to_rgb
+
+        self.failing_color = failing_color
+        self.failing_rgb = [int(x * 255) for x in to_rgb(self.failing_color)]
+
+        super().__init__(*args, **kwargs)
+
+        self.default_complete_style = self.complete_style
+        self.default_finished_style = self.finished_style
+
+    def callbacks(self, task: Task):
+        """Update bar color based on failure state."""
+        if task.fields["failing"]:
+            self.complete_style = Style.parse("rgb({},{},{})".format(*self.failing_rgb))
+            self.finished_style = Style.parse("rgb({},{},{})".format(*self.failing_rgb))
+        else:
+            # Recovered from failing
+            self.complete_style = self.default_complete_style
+            self.finished_style = self.default_finished_style
+
+
+class RichProgressBackend:
+    """Rich-based progress bar backend for terminal rendering.
+
+    This backend uses the Rich library to render progress bars in terminals
+    and Jupyter notebooks.
+    """
+
+    def __init__(
+        self,
+        chains: int,
+        total_draws: int,
+        combined: bool,
+        full_stats: bool,
+        progress_columns: list,
+        progress_stats: dict[str, list[Any]],
+        theme: Theme | None = None,
+        disable: bool = False,
+    ):
+        """Initialize the Rich progress backend.
+
+        Parameters
+        ----------
+        chains : int
+            Number of chains being sampled
+        total_draws : int
+            Total number of draws per chain (including tuning)
+        combined : bool
+            Whether to show a single combined progress bar
+        full_stats : bool
+            Whether to show detailed statistics
+        progress_columns : list
+            Rich column definitions from step method
+        progress_stats : dict
+            Initial values for statistics by chain
+        theme : Theme, optional
+            Rich theme for styling
+        disable : bool
+            Whether to disable the progress bar
+        """
+        self.chains = chains
+        self.total_draws = total_draws
+        self.combined = combined
+        self.full_stats = full_stats
+        self.progress_stats = progress_stats
+        self._disable = disable
+
+        if theme is None:
+            theme = default_progress_theme
+
+        # Build progress bar columns
+        self._progress = self._create_progress_bar(
+            progress_columns=progress_columns,
+            theme=theme,
+            disable=disable,
+        )
+        self._tasks: list[TaskID | None] = []
+
+    @property
+    def is_enabled(self) -> bool:
+        """Whether the progress bar is enabled."""
+        return not self._disable
+
+    def _create_progress_bar(
+        self,
+        progress_columns: list,
+        theme: Theme,
+        disable: bool,
+    ) -> CustomProgress:
+        """Create the Rich progress bar with appropriate columns."""
+        columns = [TextColumn("{task.fields[draws]}", table_column=Column("Draws", ratio=1))]
+
+        if self.full_stats:
+            columns += progress_columns
+
+        columns += [
+            TextColumn(
+                "{task.fields[sampling_speed]:0.2f} {task.fields[speed_unit]}",
+                table_column=Column("Sampling Speed", ratio=1),
+            ),
+            TimeElapsedColumn(table_column=Column("Elapsed", ratio=1)),
+            TimeRemainingColumn(table_column=Column("Remaining", ratio=1)),
+        ]
+
+        return CustomProgress(
+            RecolorOnFailureBarColumn(
+                table_column=Column("Progress", ratio=2),
+                failing_color="tab:red",
+                complete_style=Style.parse("rgb(31,119,180)"),  # tab:blue
+                finished_style=Style.parse("rgb(31,119,180)"),  # tab:blue
+            ),
+            *columns,
+            console=Console(theme=theme),
+            disable=disable,
+            include_headers=True,
+        )
+
+    def __enter__(self) -> RichProgressBackend:
+        """Enter the context manager and initialize tasks."""
+        self._progress.__enter__()
+        self._initialize_tasks()
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb) -> bool:
+        """Exit the context manager."""
+        result = self._progress.__exit__(exc_type, exc_val, exc_tb)
+        return result if result is not None else False
+
+    def _initialize_tasks(self) -> None:
+        """Initialize progress bar tasks for all chains."""
+        if self.combined:
+            self._tasks = [
+                self._progress.add_task(
+                    "Sampling",
+                    completed=0,
+                    draws=0,
+                    total=self.total_draws * self.chains - 1,
+                    chain_idx=0,
+                    sampling_speed=0,
+                    speed_unit="draws/s",
+                    failing=False,
+                    **{stat: value[0] for stat, value in self.progress_stats.items()},
+                )
+            ]
+        else:
+            self._tasks = [
+                self._progress.add_task(
+                    "Sampling",
+                    completed=0,
+                    draws=0,
+                    total=self.total_draws - 1,
+                    chain_idx=chain_idx,
+                    sampling_speed=0,
+                    speed_unit="draws/s",
+                    failing=False,
+                    **{stat: value[chain_idx] for stat, value in self.progress_stats.items()},
+                )
+                for chain_idx in range(self.chains)
+            ]
+
+    def update(
+        self,
+        chain_idx: int,
+        draw: int,
+        failing: bool,
+        stats: dict[str, Any],
+        is_last: bool,
+    ) -> None:
+        """Update progress bar for a specific chain.
+
+        Parameters
+        ----------
+        chain_idx : int
+            Index of the chain being updated
+        draw : int
+            Current draw number
+        failing : bool
+            Whether the chain has encountered failures
+        stats : dict
+            Statistics to display
+        is_last : bool
+            Whether this is the final update
+        """
+        if not self.is_enabled:
+            return
+
+        task_id = self._tasks[chain_idx]
+        if task_id is None:
+            return
+
+        elapsed = self._progress.tasks[chain_idx].elapsed
+        speed, unit = compute_draw_speed(elapsed, draw)
+
+        self._progress.update(
+            task_id,
+            completed=draw,
+            draws=draw,
+            sampling_speed=speed,
+            speed_unit=unit,
+            failing=failing,
+            **stats,
+        )
+
+        if is_last:
+            self._progress.update(
+                task_id,
+                draws=draw + 1 if not self.combined else draw,
+                failing=failing,
+                **stats,
+                refresh=True,
+            )
